refactor: RecruitmentCacheService를 Repository로 변경한다.
